Minimal Working example for logging message and recieving them

+++ src/Logger.hpp
@@ -7,11 +7,11 @@ class Logger
 {
 private:
     Server &m_server;
-    uint8_t m_logger_id;
+    int m_logger_id;
 
 public:
     Logger() = delete;
-    Logger(Server &server, uint8_t logger_id, const std::string &title);
+    Logger(Server &server, int logger_id, const std::string &title);
 
     void log(const std::string &message);
 };
